requirements.txt
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
pytesseract==0.3.10
